feat: AudioButton Entity完全削除とPlainObject移行
- AudioButton Entityクラスとすべての関連Value Objectsを削除
- AudioButton関連のインポートをPlainObjectとCompatに移行
- CreateAudioButtonInputとUpdateAudioButtonInput型を追加
- AudioButtonQuery型をアプリケーション側で定義
- テストをPlainObject使用に更新（674件のテストがパス）
- Lint警告を修正

削除されたファイル:
- packages/shared-types/src/entities/audio-button.ts
- packages/shared-types/src/entities/__tests__/audio-button.test.ts
- packages/shared-types/src/value-objects/audio-button/ (全ディレクトリ)

移行の詳細はADR-004を参照
